refactor(workspace): extract quspin-basis + OperatorDispatch trait (PR 4/7)
Fourth of seven PRs splitting quspin-core. Extracts all basis types
into a new quspin-basis crate, and introduces the `OperatorDispatch`
extension trait to keep the basis-dependent dispatch methods on
`*OperatorInner` without putting the enums themselves in quspin-core.

## Structural changes

1. `*OperatorInner` enum definitions (BondOperatorInner, SpinOperatorInner,
   etc.) move from quspin-core/src/operator/*/dispatch.rs to
   quspin-operator/src/*/dispatch.rs. Only the non-basis methods
   (max_site, num_cindices, lhss) stay inherent on the enum.

2. `apply_and_project_to` / `apply` inherent methods become an
   `OperatorDispatch` extension trait defined in quspin-core (to be
   moved to quspin-matrix in PR 5). Each `*OperatorInner` gets an impl
   block in its dispatch.rs that calls the free function in
   operator/apply.rs.

3. All basis files (bfs, boson, dispatch, expand, fermion, generic,
   lattice, orbit, seed, space, spin, sym, traits) move verbatim to
   quspin-basis. quspin-basis depends on quspin-operator for
   `*OperatorInner` types used by the `build_*` methods, and on
   quspin-bitbasis + quspin-types for primitive support.

4. quspin-core/src/basis/mod.rs becomes a thin
   `pub use quspin_basis::*` shim.

## Deviation from spec §4 Move 2

The spec's target is a fully acyclic DAG where quspin-basis and
quspin-operator compile in parallel. Achieving that would require
removing the `build_*` inherent methods from the basis types and
exposing them as free functions in quspin-matrix, which forces
quspin-py to call `spin_build_spin(&mut basis, &op.inner, seeds)`
instead of `basis.build_spin(&op.inner, seeds)`.

To preserve the `basis.build_*` method syntax (and minimize quspin-py
churn), quspin-basis keeps the `build_*` methods inherent and takes
a dep on quspin-operator. This means quspin-operator and quspin-basis
do NOT compile in parallel, but they form a clean DAG with no cycle.

## quspin-py changes

Five files gain a single `use quspin_core::OperatorDispatch;` line so
method-call syntax resolves to the new trait impls:
- operator/{bond,boson,fermion,monomial,pauli}.rs

The Cargo.toml and all other bindings are untouched.
