+++ crates/quspin-basis/Cargo.toml
@@ -1,22 +1,17 @@
 [package]
-name = "quspin-core"
+name = "quspin-basis"
 version.workspace = true
 edition.workspace = true
 
 [features]
-large-int = ["quspin-bitbasis/large-int", "quspin-basis/large-int"]
+large-int = ["quspin-bitbasis/large-int"]
 
 [dependencies]
 quspin-types = { path = "../quspin-types" }
 quspin-bitbasis = { path = "../quspin-bitbasis" }
 quspin-operator = { path = "../quspin-operator" }
-quspin-basis = { path = "../quspin-basis" }
-num-complex = { workspace = true }
 ndarray = { workspace = true }
+num-complex = { workspace = true }
 rayon = { workspace = true }
 ruint = { workspace = true }
 smallvec = { workspace = true }
-thiserror = { workspace = true }
-bytemuck = { workspace = true }
-ode_solvers = { workspace = true }
-nalgebra = { workspace = true }
