+++ crates/quspin-basis/src/bfs.rs
