+++ crates/quspin-basis/src/lattice.rs
