+++ crates/quspin-basis/src/seed.rs
