+++ crates/quspin-operator/src/bond/mod.rs
@@ -1,3 +1,5 @@
+pub mod dispatch;
 pub mod operator;
 
+pub use dispatch::BondOperatorInner;
 pub use operator::{BondOperator, BondTerm};
