+++ crates/quspin-operator/src/boson/mod.rs
@@ -1,3 +1,5 @@
+pub mod dispatch;
 pub mod operator;
 
+pub use dispatch::BosonOperatorInner;
 pub use operator::{BosonOp, BosonOpEntry, BosonOperator};
