+++ crates/quspin-operator/src/fermion/mod.rs
@@ -1,3 +1,5 @@
+pub mod dispatch;
 pub mod operator;
 
+pub use dispatch::FermionOperatorInner;
 pub use operator::{FermionOp, FermionOpEntry, FermionOperator};
