+++ crates/quspin-operator/src/monomial/mod.rs
@@ -1,3 +1,5 @@
+pub mod dispatch;
 pub mod operator;
 
+pub use dispatch::MonomialOperatorInner;
 pub use operator::{MonomialOperator, MonomialTerm};
