+++ crates/quspin-operator/src/pauli/mod.rs
@@ -1,3 +1,5 @@
+pub mod dispatch;
 pub mod operator;
 
+pub use dispatch::HardcoreOperatorInner;
 pub use operator::{HardcoreOp, HardcoreOperator, OpEntry};
