+++ crates/quspin-operator/src/spin/mod.rs
@@ -1,3 +1,5 @@
+pub mod dispatch;
 pub mod operator;
 
+pub use dispatch::SpinOperatorInner;
 pub use operator::{SpinOp, SpinOpEntry, SpinOperator};
