+++ crates/quspin-py/src/operator/boson.rs
@@ -5,6 +5,7 @@ use crate::operator::{
 };
 use numpy::{Complex64, PyArray1};
 use pyo3::prelude::*;
+use quspin_core::OperatorDispatch;
 use quspin_core::operator::boson::{BosonOp, BosonOpEntry, BosonOperator, BosonOperatorInner};
 
 /// Python-facing bosonic operator (truncated harmonic oscillator).
